Audit other DB conn tests.
TST: Audit test_set_telescope_pointing
and split DB and non-DB tests
and attempt to minimize DB conn in the DB ones within reason
