+++ jwst/regtest/lib/test_v1_calculate.py
@@ -1,15 +1,29 @@
-"""Test module v1_calculate"""
+"""Test module v1_calculate."""
 
-import numpy as np
 import pytest
-from astropy.table import Table
-from astropy.time import Time
-from astropy.utils.data import get_pkg_data_filename
-from stdatamodels.jwst.datamodels import ImageModel
 
-import jwst.lib.set_telescope_pointing as stp
-import jwst.lib.v1_calculate as v1c
-from jwst.lib import engdb_mast
+pytest.importorskip("pysiaf")
+
+import numpy as np  # noqa: E402
+from astropy.table import Table  # noqa: E402
+from astropy.time import Time  # noqa: E402
+from astropy.utils.data import get_pkg_data_filename  # noqa: E402
+from stdatamodels.jwst.datamodels import ImageModel  # noqa: E402
+
+import jwst.lib.set_telescope_pointing as stp  # noqa: E402
+import jwst.lib.v1_calculate as v1c  # noqa: E402
+from jwst.lib import engdb_mast  # noqa: E402
+
+# Mark all tests in this module as slow due to possible remote DB connection
+pytestmark = [pytest.mark.slow]
+
+# The tests here need MAST DB to be available.
+try:
+    engdb_mast.EngdbMast(base_url=engdb_mast.MAST_BASE_URL)
+except RuntimeError as exception:
+    pytest.skip(
+        f"Live MAST Engineering Service not available: {exception}", allow_module_level=True
+    )
 
 # Engineering parameters
 # Time range corresponds to OTE-1 exposure jw01134001037_03107_00001_nrcb1_uncal.fits
@@ -19,19 +33,14 @@ MAST_GOOD_ENDTIME = Time("59612.93500967592", format="mjd")
 
 
 def test_from_models_mast(tmp_path):
-    """Test v1_calculate_from_models for basic running"""
+    """Test v1_calculate_from_models for basic running."""
     model = ImageModel()
     model.meta.exposure.start_time = MAST_GOOD_STARTTIME.mjd
     model.meta.exposure.end_time = MAST_GOOD_ENDTIME.mjd
 
-    try:
-        v1_table = v1c.v1_calculate_from_models(
-            [model], method=stp.Methods.COARSE_TR_202111, engdb_url=engdb_mast.MAST_BASE_URL
-        )
-    except ValueError as exception:
-        pytest.xfail(
-            f"MAST engineering database not available, possibly no token specified: {exception}"
-        )
+    v1_table = v1c.v1_calculate_from_models(
+        [model], method=stp.Methods.COARSE_TR_202111, engdb_url=engdb_mast.MAST_BASE_URL
+    )
     v1_formatted = v1c.simplify_table(v1_table)
 
     # Save for post-test examination
@@ -46,18 +55,13 @@ def test_from_models_mast(tmp_path):
 
 
 def test_over_time_mast(tmp_path):
-    """Test v1_calculate_over_time for basic running"""
-    try:
-        v1_table = v1c.v1_calculate_over_time(
-            MAST_GOOD_STARTTIME.mjd,
-            MAST_GOOD_ENDTIME.mjd,
-            method=stp.Methods.COARSE_TR_202111,
-            engdb_url=engdb_mast.MAST_BASE_URL,
-        )
-    except ValueError as exception:
-        pytest.xfail(
-            f"MAST engineering database not available, possibly no token specified: {exception}"
-        )
+    """Test v1_calculate_over_time for basic running."""
+    v1_table = v1c.v1_calculate_over_time(
+        MAST_GOOD_STARTTIME.mjd,
+        MAST_GOOD_ENDTIME.mjd,
+        method=stp.Methods.COARSE_TR_202111,
+        engdb_url=engdb_mast.MAST_BASE_URL,
+    )
     v1_formatted = v1c.simplify_table(v1_table)
 
     # Save for post-test examination
@@ -74,13 +78,10 @@ def test_over_time_mast(tmp_path):
 # ######################
 # Fixtures and utilities
 # ######################
-@pytest.fixture
-def engdb_service():
-    yield engdb_mast.EngdbMast()
 
 
 def v1_compare_simplified_tables(a, b, rtol=1e-05):
-    """Calculate diff between tables generated by v1_calculate"""
+    """Calculate diff between tables generated by v1_calculate."""
     errors = []
     if len(a) != len(b):
         errors.append(f"len(a)={len(a)} not equal to len(b)={len(b)}")
